fix(diagnostics): move slice-3 additions into their own modules for the max-lines gate
producer/mod.rs (641) and supervisor/lifecycle.rs (674) both breached the
600-line repo shape cap. Client: the generation availability transitions
and their loud logging move to producer/delivery_log.rs (kept out of the
mod.rs regions slice 2b reshuffles). Desktop: the death-certificate type
and the supervisor's certificate-assembly helpers move to
supervisor/death_certificate.rs. No behavior change; negative controls
re-verified after the move.
