+++ apps/desktop/src-tauri/src/diagnostics_collector/supervisor/lifecycle.rs
@@ -3,10 +3,9 @@ use std::sync::Arc;
 use std::time::Instant;
 
 use proliferate_diagnostics_protocol::v1::limits::CURRENT_SCHEMA_VERSION;
-use proliferate_diagnostics_protocol::v1::types::{
-    ArgumentValueV1, HealthStatusV1, PrivacyClassificationV1, TerminalOutcomeV1, TypedArgumentV1,
-};
+use proliferate_diagnostics_protocol::v1::types::{HealthStatusV1, TerminalOutcomeV1};
 
+use super::death_certificate::CollectorDeathCertificate;
 use super::{
     accept_restart_budget_at, classification_for_launch_error, reset_restart_budget_after_health,
     retryable, CollectorClientError, CollectorLaunchError, CollectorLaunchErrorKind,
@@ -15,65 +14,6 @@ use super::{
     SupervisorUnavailable, AUTOMATIC_RESTART_DELAYS, HEALTH_FAILURE_THRESHOLD,
     PRODUCER_DRAIN_TIMEOUT, STEADY_HEALTH_INTERVAL,
 };
-
-/// Death certificate for a failed collector generation: what the monitor (or
-/// the stop path) observed at the moment the child was declared gone. The
-/// exit status is present only when the child was actually seen exited;
-/// inspection failures, health losses, and latched classifications carry none.
-#[derive(Debug, Clone, Copy)]
-pub(super) struct CollectorDeathCertificate {
-    trigger: &'static str,
-    exit_status: Option<std::process::ExitStatus>,
-}
-
-impl CollectorDeathCertificate {
-    pub(super) fn new(
-        trigger: &'static str,
-        exit_status: Option<std::process::ExitStatus>,
-    ) -> Self {
-        Self {
-            trigger,
-            exit_status,
-        }
-    }
-
-    pub(super) fn arguments(&self, restart_count: u64) -> Vec<TypedArgumentV1> {
-        let mut arguments = vec![
-            operational_argument("trigger", ArgumentValueV1::String(self.trigger.to_owned())),
-            operational_argument(
-                "restart_count",
-                ArgumentValueV1::Integer(i64::try_from(restart_count).unwrap_or(i64::MAX)),
-            ),
-        ];
-        if let Some(status) = self.exit_status {
-            if let Some(code) = status.code() {
-                arguments.push(operational_argument(
-                    "exit_code",
-                    ArgumentValueV1::Integer(i64::from(code)),
-                ));
-            }
-            #[cfg(unix)]
-            {
-                use std::os::unix::process::ExitStatusExt;
-                if let Some(signal) = status.signal() {
-                    arguments.push(operational_argument(
-                        "signal",
-                        ArgumentValueV1::Integer(i64::from(signal)),
-                    ));
-                }
-            }
-        }
-        arguments
-    }
-}
-
-fn operational_argument(name: &str, value: ArgumentValueV1) -> TypedArgumentV1 {
-    TypedArgumentV1 {
-        name: name.to_owned(),
-        privacy: PrivacyClassificationV1::Operational,
-        value,
-    }
-}
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(super) enum SteadyHealthAssessment {
@@ -188,15 +128,7 @@ impl DiagnosticsCollectorSupervisor {
             Ok(Some(status)) => {
                 process.finish_reaped();
                 self.invalidate_generation();
-                let restart_count = self
-                    .inner
-                    .lock()
-                    .map(|inner| inner.restart_count)
-                    .unwrap_or_default();
-                stop_operation.append_arguments(
-                    CollectorDeathCertificate::new("child_exited", Some(status))
-                        .arguments(restart_count),
-                );
+                stop_operation.append_arguments(self.child_exited_arguments(status));
                 stop_operation.terminal(TerminalOutcomeV1::Failed, Some("child_exited"));
                 self.publish_state(DesktopDiagnosticsSupervisorStateV1::Stopped { orderly: false });
                 return Ok(());
@@ -259,23 +191,7 @@ impl DiagnosticsCollectorSupervisor {
                 self.publish_degraded("restart_exhausted", true);
                 return false;
             }
-            let restart_operation = match certificate {
-                Some(certificate) => {
-                    // `accept_restart_budget` already counted this attempt, so
-                    // the certificate's restart_count matches the Ready state
-                    // this attempt would publish.
-                    let restart_count = self
-                        .inner
-                        .lock()
-                        .map(|inner| inner.restart_count)
-                        .unwrap_or_default();
-                    self.producer.begin_lifecycle_with_arguments(
-                        "desktop.collector.restart",
-                        certificate.arguments(restart_count),
-                    )
-                }
-                None => self.producer.begin_lifecycle("desktop.collector.restart"),
-            };
+            let restart_operation = self.begin_restart_lifecycle(certificate);
             self.publish_starting(CollectorLaunchKindV1::AutomaticRestart, (index + 1) as u8);
             let launcher = match &self.launcher {
                 Ok(launcher) => launcher,
